Refactor NobackupContextWrapper
 - Make internal
 - Move to internal package
 - Remove val for context
